Master

See merge request CrowdRobot/crowdrobot!33

+++ helper_funcs/misc.py
@@ -28,25 +28,28 @@ def delete_messages(update, context, message_from_update=False):
                                        update.effective_message.message_id)
         except TelegramError:
             pass
-    if 'to_delete' in context.user_data:
-        for msg in context.user_data['to_delete']:
-            if type(msg) is list:
-                for ms in msg:
-                    try:
-                        if ms.message_id != update.effective_message.message_id:
-                            context.bot.delete_message(
-                                update.effective_chat.id, ms.message_id)
-                    except TelegramError:
-                        continue
-            else:
-                if update.effective_message and msg:
-                    try:
-                        if msg.message_id != update.effective_message.message_id:
-                            context.bot.delete_message(
-                                update.effective_chat.id, msg.message_id)
-                    except TelegramError:
-                        continue
-        context.user_data['to_delete'] = list()
+    if context.user_data:
+        if 'to_delete' in context.user_data:
+            for msg in context.user_data['to_delete']:
+                if type(msg) is list:
+                    for ms in msg:
+                        try:
+                            if ms.message_id != update.effective_message.message_id:
+                                context.bot.delete_message(
+                                    update.effective_chat.id, ms.message_id)
+                        except TelegramError:
+                            continue
+                else:
+                    if update.effective_message and msg:
+                        try:
+                            if msg.message_id != update.effective_message.message_id:
+                                context.bot.delete_message(
+                                    update.effective_chat.id, msg.message_id)
+                        except TelegramError:
+                            continue
+            context.user_data['to_delete'] = list()
+        else:
+            context.user_data['to_delete'] = list()
     else:
         context.user_data['to_delete'] = list()
